+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed Node.js warnings such as "Warning: Closing file descriptor # on garbage collection", which is deprecated in Node.js 22 and will result in an error being thrown in the future
+
 ## 0.1.4 (2024-10-30)
 
 - Added `repo.currentBranch()` method